pyproject.toml
[tool.poetry]
name = "zno-explorer-backend"
version = "0.1.0"
description = ""
authors = ["Suhas N <suhas.nandibhatla@gmail.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "pytorch_cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.2"
numpy = "^2.1.1"
pillow = "^10.4.0"
orjson = "^3.10.7"
opencv-python-headless = "^4.10.0.84"
torch = "^2.4.1+cpu"
torchvision = "0.19.1+cpu"
scikit-image = "^0.24.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"